refactor: move visualize commands
