+++ src/commands/index.ts
@@ -1,0 +1,2 @@
+export { visualize } from './visualize';
+export { visualizeProject } from './visualizeProject';
